refactored transactions table to be more reusable. created components for pagination and limit selection
